+++ lib/ap/mips/vlrt.c
@@ -7,7 +7,7 @@
  * in the LICENSE file.
  */
 
-typedef	unsigned long	ulong;
+typedef	uint32_t	ulong;
 typedef	unsigned int	uint;
 typedef	unsigned short	ushort;
 typedef	unsigned char	uchar;
@@ -22,8 +22,8 @@ struct	Vlong
 	{
 		struct
 		{
-			unsigned long	hi;
-			unsigned long	lo;
+			uint32_t	hi;
+			uint32_t	lo;
 		};
 		struct
 		{
@@ -40,7 +40,7 @@ void	abort(void);
 void
 _addv(Vlong *r, Vlong a, Vlong b)
 {
-	unsigned long lo, hi;
+	uint32_t lo, hi;
 
 	lo = a.lo + b.lo;
 	hi = a.hi + b.hi;
@@ -53,7 +53,7 @@ _addv(Vlong *r, Vlong a, Vlong b)
 void
 _subv(Vlong *r, Vlong a, Vlong b)
 {
-	unsigned long lo, hi;
+	uint32_t lo, hi;
 
 	lo = a.lo - b.lo;
 	hi = a.hi - b.hi;
@@ -67,7 +67,7 @@ void
 _d2v(Vlong *y, double d)
 {
 	union { double d; struct Vlong; } x;
-	unsigned long xhi, xlo, ylo, yhi;
+	uint32_t xhi, xlo, ylo, yhi;
 	int sh;
 
 	x.d = d;
@@ -149,7 +149,7 @@ _v2f(Vlong x)
 static void
 dodiv(Vlong num, Vlong den, Vlong *qp, Vlong *rp)
 {
-	unsigned long numlo, numhi, denhi, denlo, quohi, quolo, t;
+	uint32_t numlo, numhi, denhi, denlo, quohi, quolo, t;
 	int i;
 
 	numhi = num.hi;
@@ -314,7 +314,7 @@ _rshav(Vlong *r, Vlong a, int b)
 void
 _rshlv(Vlong *r, Vlong a, int b)
 {
-	unsigned long t;
+	uint32_t t;
 
 	t = a.hi;
 	if(b >= 32) {
@@ -339,7 +339,7 @@ _rshlv(Vlong *r, Vlong a, int b)
 void
 _lshv(Vlong *r, Vlong a, int b)
 {
-	unsigned long t;
+	uint32_t t;
 
 	t = a.lo;
 	if(b >= 32) {
@@ -488,10 +488,10 @@ _vasop(Vlong *ret, void *lv, void fn(Vlong*, Vlong, Vlong), int type, Vlong rv)
 		break;
 
 	case 6:	/* ulong */
-		t.lo = *(unsigned long*)lv;
+		t.lo = *(uint32_t*)lv;
 		t.hi = 0;
 		fn(&u, t, rv);
-		*(unsigned long*)lv = u.lo;
+		*(uint32_t*)lv = u.lo;
 		break;
 
 	case 7:	/* vlong */
@@ -508,7 +508,7 @@ _p2v(Vlong *ret, void *p)
 {
 	long t;
 
-	t = (unsigned long)p;
+	t = (uint32_t)p;
 	ret->lo = t;
 	ret->hi = 0;
 }
@@ -524,7 +524,7 @@ _sl2v(Vlong *ret, long sl)
 }
 
 void
-_ul2v(Vlong *ret, unsigned long ul)
+_ul2v(Vlong *ret, uint32_t ul)
 {
 	long t;
 
@@ -564,7 +564,7 @@ _sh2v(Vlong *ret, long sh)
 }
 
 void
-_uh2v(Vlong *ret, unsigned long ul)
+_uh2v(Vlong *ret, uint32_t ul)
 {
 	long t;
 
@@ -584,7 +584,7 @@ _sc2v(Vlong *ret, long uc)
 }
 
 void
-_uc2v(Vlong *ret, unsigned long ul)
+_uc2v(Vlong *ret, uint32_t ul)
 {
 	long t;
