+++ lib/ap/stdio/vfwprintf.c
@@ -128,7 +128,7 @@ static void pop_arg(union arg *arg, int type, va_list *ap)
 	break; case UINT:	arg->i = va_arg(*ap, unsigned int);
 #ifndef LONG_IS_INT
 	break; case LONG:	arg->i = va_arg(*ap, long);
-	break; case ULONG:	arg->i = va_arg(*ap, unsigned long);
+	break; case ULONG:	arg->i = va_arg(*ap, uint32_t);
 #endif
 	break; case ULLONG:	arg->i = va_arg(*ap, unsigned long long);
 	break; case SHORT:	arg->i = (short)va_arg(*ap, int);
